Update genomebam input
Add new function to define input for genomebam. Input should only be correct when activated in config file, with the input unable to be created if not activated. In this way, the pipeline should use the kallisto_quant rule when genomebam deactivated in the config file as the there is no rule to create the dummy input