Add diffexp suffix parameter
Adds new parameter diffexp. This suffix parameter defines the string used to iterate over the so$target_mapping data frame to update the annotation info for entries labelled with the suffix. This corrects the custom transcripts generated by TRAWLING prior to producing the output rds object in the sleut_init rule.