Overhaul script to support TRAWLING entries
A large number of changes were made to the code to support TRAWLING entries
- Now containes two instances for generating plots; first to visualise the top n transcripts, followed by a second instance that then visualises transcripts for genes of interest

- The original code combined the top n genes with genes of interest into a single list to generate the plots in one go. The two main issues in the way it was coded were 1) only a single transcript was visualised for each gene of interest in the param file and 2) only canonical transcripts were being visualised. These have been addressed

- Both canonical and non-canonical transcripts within the top n are visualised, with all transcripts for a gene of interest also now visualised provided there are no NA values in the data (excepting canonical status), regardless of significance

- Replaces ":" with "_" from transcript names when saving, which caused compatibility issues (files cannot be saved with ":" special characters)


Of note, in the original code a list of genes (including genes of interest) is created, with an additional entry created for "Custom" transcript <add_row(ext_gene = "Custom") >
This is redundant though, as in the sleuth_init rule, the R script automatically assigns custom transcripts with NA values for ens_gene and canonical columns resulting in them getting dropped due to a <drop_na()> command. While this commit removes this code, functionality for custom transcripts still remains missing. This comment here highlights this in case we wish to fix this functionality in the future.